+++ lib/src/flutter/extensions/num_extension.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 ///Num extension to extend num functionality
 extension NumExt on num {
   /// Creates a fixed square sized box.
-  SizedBox box([Widget? child]) =>
+  SizedBox sq([Widget? child]) =>
       SizedBox(height: toDouble(), width: toDouble(), child: child);
 
   /// Create Fixed size width box